Move extension definitions to their own file
